Modify to work as a compiled script
Required many changes to how files are handled.

Created a Resouce.ahk module which is #Include'd in PACS Assistant, to bring in file resources as needed for the compiled version to work.

Created a Compile.ahk module which is a standalone script that generates the Resource.ahk module.

To create a compiled version, first run Compile.ahk to update Resource.ahk.

Then use ahk2exe to make a compiled exe version.